requirements.txt
google-generativeai==0.8.3
streamlit
pydanti
cryptography
Pillow